refactor(pipeline): split odpt/ into odpt-common/ and odpt-train/
The dir name odpt/ implied a generic ODPT scope, but every module
under it depended on TRAIN-only types (OdptRailway, OdptStation,
OdptStationTimetable). This was inconsistent with the rest of the
repo, which already uses odpt-train as the canonical scope name
(build-from-odpt-train.ts, load-odpt-train-sources.ts,
types/odpt-train.ts, npm script pipeline:build:v2-odpt-train).

ODPT API spec v4.15 categorises classes as 共通 (Calendar / Operator),
鉄道, バス, 航空機. Reflect this in the dir layout:

- odpt-common/ holds modules that depend on no TRAIN types and would
  apply unchanged to a future BUS / AIRPLANE pipeline:
  - build-agency.ts (only reads Provider config)
  - build-feed-info.ts (only reads Provider + issued date)
- odpt-train/ holds everything else (inherited from the old odpt/),
  including the Yurikamome-tuned trip-identity heuristic.

build-calendar.ts conceptually belongs in common (only reads the
odpt:calendar field) but its current signature accepts
OdptStationTimetable[]. Kept in odpt-train/ for now under a
strict-reading interpretation; can be lifted when BUS support lands.

Pure structural refactor (git mv only): yurimo data.json + all 59
data-v2 build artifacts byte-identical pre/post (md5 diff exit=0).
